refactor(utils): split Cognito lookups out of request, add subpath exports
A bundled Lambda needing only the request wrapper pulled in the whole barrel:
@aws-sdk/client-cognito-identity-provider via request.ts, which held both
getContext and the Cognito user lookups, plus client-dynamodb and
util-dynamodb via repository.ts.

getContext only parses the incoming event, so the Cognito-backed functions
(isUserInGroup, getUserById, getUser, _FindUser) move to ./users. The barrel
re-exports them, so the public API is unchanged.

Also adds an exports map giving the pieces clean specifiers
(@lambada/utils/handler, /cors, /request, /users, /repository, /error).

The ./dist/* patterns are deliberate, not leftovers: six services already
deep-import extensionless paths -- chat, compliance, notification, orderbook,
pricing and tatum-provider -- and adding `exports` otherwise gates them off.
Both "./dist/*.js" and "./dist/*" are declared because Node does not retry an
array fallback target when the file is missing, so a single pattern breaks one
of the two forms. Verified: every specifier those six use resolves under both
require() and tsc with moduleResolution node16.

Bundling the wrapper alone: 12,605 -> 3,854 bytes, with no AWS SDK require.
