+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
 export * from './repository'
 export * from './request'
+export * from './users'
 export * from './cors'
 export * from './handler'
 export * from './data'
